Обновить логику хранения рейтингов

+++ src/database/migrations/2018_09_26_041303_create_ratings_table.php
@@ -16,7 +16,7 @@ class CreateRatingsTable extends Migration
         Schema::create('ratings', function (Blueprint $table) {
             $table->unsignedInteger('book_id');
             $table->unsignedInteger('user_id');
-            $table->unsignedInteger('rating');
+            $table->float('rating');
             $table->primary(['book_id', 'user_id']);
             $table->foreign('book_id')->references('id')->on('books')->onDelete('CASCADE');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('CASCADE');
